# packages/vibeship-mind/vibeship_mind-5.0.1.tar.gz/vibeship_mind-5.0.1/docs/specs/openapi.yaml
openapi: 3.1.0
info:
  title: Mind v5 API
  description: |
    Decision intelligence API for AI agents. Mind v5 provides hierarchical temporal memory,
    outcome-weighted retrieval, causal reasoning, and collective intelligence through
    federated pattern learning.

    ## Core Concepts

    - **Memories**: Hierarchical temporal memories with outcome-weighted salience
    - **Decisions**: Tracked decision traces with context snapshots
    - **Outcomes**: Feedback signals that adjust memory salience for future retrieval
    - **Context**: Fused retrieval from vector, graph, and keyword sources

    ## Authentication

    All endpoints require Bearer JWT authentication. Tokens are scoped to a single user
    and include the user_id claim.

    ## Rate Limits

    - Standard: 1000 requests/minute
    - Burst: 10000 requests/minute (short bursts)
    - Memory retrieval: 100 requests/second per user

    ## Versioning

    This API follows semantic versioning. Breaking changes require a major version bump.
    The current version is v1. Use the `Accept-Version` header for version negotiation.
  version: 1.0.0
  contact:
    name: Mind v5 Team
    email: mind-api@example.com
    url: https://docs.mind-v5.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.mind-v5.dev/v1
    description: Production server
  - url: https://staging-api.mind-v5.dev/v1
    description: Staging server
...
security:
  - BearerAuth: []
...
paths:
  /memories/retrieve:
    post:
      operationId: retrieveMemories
      summary: Retrieve context with outcome weighting
      description: |
        Retrieves relevant memories for a given query using multi-source fusion:
        - Vector similarity (Qdrant)
        - Graph traversal (FalkorDB, 2-hop)
        - Keyword/BM25 (PostgreSQL full-text)
        - Temporal recency (decay function)
        - Outcome-weighted salience (learned from feedback)

        Results are fused using Reciprocal Rank Fusion (RRF) and optionally
        reranked with a cross-encoder for maximum relevance.

        **SLO**: 200ms p99 latency
      tags:
        - Memories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetrievalRequest'
            examples:
              basic:
                summary: Basic retrieval
                value:
                  query: "What are the user's communication preferences?"
                  limit: 10
              with_levels:
                summary: Filtered by temporal levels
                value:
                  query: "What projects is the user working on?"
                  limit: 20
                  temporal_levels:
                    - identity
                    - seasonal
                  include_causal: true
              with_session:
                summary: Session-scoped retrieval
                value:
                  query: "What did we discuss earlier?"
                  limit: 5
                  session_id: "550e8400-e29b-41d4-a716-446655440000"
                  temporal_levels:
                    - immediate
                    - situational
      responses:
        '200':
          description: Successfully retrieved memories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrievalResponse'
              examples:
                success:
                  summary: Successful retrieval
                  value:
                    trace_id: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
                    memories:
                      - memory_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                        content: "User prefers concise, bullet-point responses"
                        temporal_level: identity
                        salience: 0.92
                        outcome_score: 0.15
                        recency_score: 0.3
                        fusion_rank: 1
                      - memory_id: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
                        content: "Currently working on Mind v5 architecture"
                        temporal_level: seasonal
                        salience: 0.85
                        outcome_score: 0.08
                        recency_score: 0.7
                        fusion_rank: 2
                    causal_edges: []
                    retrieval_latency_ms: 47.3
                    fusion_method: "rrf"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /memories:
    post:
      operationId: createMemory
      summary: Store a new memory
      description: |
        Creates a new memory in the hierarchical temporal memory system.

        Memories are automatically:
        - Embedded using text-embedding-3-small (1536 dimensions)
        - Indexed in Qdrant for vector search
        - Linked to entities via FalkorDB graph
        - Assigned initial salience based on content type

        **Temporal Levels**:
        - `immediate`: Current session, working memory (hours)
        - `situational`: Active tasks, recent events (weeks)
        - `seasonal`: Projects, recurring patterns (months)
        - `identity`: Core values, stable preferences (years)

        Memories can be promoted to higher levels when patterns stabilize.
      tags:
        - Memories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMemoryRequest'
            examples:
              preference:
                summary: User preference
                value:
                  content: "User prefers dark mode and minimal UI animations"
                  content_type: preference
                  temporal_level: identity
                  metadata:
                    source: "settings_observation"
                    confidence: 0.95
              task:
                summary: Active task
                value:
                  content: "Working on implementing the causal inference engine"
                  content_type: task
                  temporal_level: situational
                  valid_until: "2025-01-15T00:00:00Z"
                  metadata:
                    project: "mind-v5"
                    priority: "high"
      responses:
        '201':
          description: Memory created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
              examples:
                created:
                  summary: Created memory
                  value:
                    memory_id: "c3d4e5f6-a7b8-9012-cdef-123456789012"
                    content: "User prefers dark mode and minimal UI animations"
                    content_type: preference
                    temporal_level: identity
                    valid_from: "2025-12-27T19:30:00Z"
                    valid_until: null
                    base_salience: 1.0
                    outcome_adjustment: 0.0
                    effective_salience: 1.0
                    retrieval_count: 0
                    decision_count: 0
                    created_at: "2025-12-27T19:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /memories/{memory_id}:
    get:
      operationId: getMemory
      summary: Get a specific memory
      description: |
        Retrieves a single memory by its ID. Includes full metadata,
        salience scores, and usage statistics.
      tags:
        - Memories
      parameters:
        - name: memory_id
          in: path
          required: true
          description: Unique identifier of the memory
          schema:
            type: string
            format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        - name: include_usage_stats
          in: query
          required: false
          description: Include detailed usage statistics
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Memory found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    patch:
      operationId: updateMemory
      summary: Update a memory
      description: |
        Updates a memory's metadata or content. Temporal level changes
        trigger re-evaluation of the memory's position in the hierarchy.

        Note: Direct salience modifications are not allowed through this
        endpoint. Salience is adjusted through the outcome observation system.
      tags:
        - Memories
      parameters:
        - name: memory_id
          in: path
          required: true
          description: Unique identifier of the memory
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemoryRequest'
      responses:
        '200':
          description: Memory updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      operationId: expireMemory
      summary: Expire a memory
      description: |
        Marks a memory as expired by setting its `valid_until` timestamp.
        Expired memories are not returned in retrieval but are preserved
        for audit and potential reactivation.

        **Note**: Memories are never permanently deleted due to event sourcing
        requirements. Use this endpoint to logically remove memories.
      tags:
        - Memories
      parameters:
        - name: memory_id
          in: path
          required: true
          description: Unique identifier of the memory
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Memory expired successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token with the following claims:
        - `sub`: User ID (UUID)
        - `iat`: Issued at timestamp
        - `exp`: Expiration timestamp
        - `scope`: Allowed operations (read, write, admin)
  schemas:
...
    TemporalLevel:
      type: string
      description: |
        Hierarchical temporal level:
        - `immediate`: Current session, working memory (hours)
        - `situational`: Active tasks, recent events (weeks)
        - `seasonal`: Projects, recurring patterns (months)
        - `identity`: Core values, stable preferences (years)
      enum:
        - immediate
        - situational
        - seasonal
        - identity
      example: identity
    CreateMemoryRequest:
      type: object
      required:
        - content
        - content_type
        - temporal_level
      properties:
        content:
          type: string
          description: The memory content
          maxLength: 10000
          example: "User prefers dark mode and minimal UI animations"
        content_type:
          type: string
          enum:
            - fact
            - preference
            - event
            - task
            - goal
            - relationship
            - skill
            - constraint
          example: preference
        temporal_level:
          $ref: '#/components/schemas/TemporalLevel'
        valid_until:
          type: string
          format: date-time
          nullable: true
          description: Optional expiration timestamp
        metadata:
          type: object
          additionalProperties: true
          description: Optional metadata
    UpdateMemoryRequest:
      type: object
      properties:
        content:
          type: string
          maxLength: 10000
          description: Updated content
        content_type:
          type: string
          enum:
            - fact
            - preference
            - event
            - task
            - goal
            - relationship
            - skill
            - constraint
        temporal_level:
          $ref: '#/components/schemas/TemporalLevel'
        valid_until:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
          additionalProperties: true

    # === Retrieval Schemas ===
    RetrievalRequest:
      type: object
      description: Request for context retrieval
      required:
        - query
      properties:
        query:
          type: string
          description: Natural language query for retrieval
          maxLength: 2000
          example: "What are the user's communication preferences?"
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
          description: Maximum number of memories to return
          example: 10
        temporal_levels:
          type: array
          items:
            $ref: '#/components/schemas/TemporalLevel'
          description: Filter to specific temporal levels
          example:
            - identity
            - seasonal
        session_id:
          type: string
          format: uuid
          description: Scope to specific session
        include_causal:
          type: boolean
          default: false
          description: Include causal edges in response
        min_salience:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          default: 0.0
          description: Minimum effective salience threshold
        rerank:
          type: boolean
          default: true
          description: Apply cross-encoder reranking
    RetrievalResponse:
      type: object
      description: Response containing retrieved memories
      required:
        - trace_id
        - memories
        - retrieval_latency_ms
        - fusion_method
      properties:
        trace_id:
          type: string
          format: uuid
          description: Trace ID for this retrieval (use in decision tracking)
          example: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
        memories:
          type: array
          items:
            $ref: '#/components/schemas/RetrievedMemory'
          description: Retrieved memories ordered by relevance
        causal_edges:
          type: array
          items:
            $ref: '#/components/schemas/CausalEdge'
          description: Related causal edges (if requested)
        retrieval_latency_ms:
          type: number
          format: float
          description: Total retrieval latency in milliseconds
          example: 47.3
        fusion_method:
          type: string
          enum:
            - rrf
            - weighted
            - simple
          description: Fusion method used
          example: rrf
    RetrievedMemory:
      type: object
      description: A memory item with retrieval scores
      required:
        - memory_id
        - content
        - temporal_level
        - salience
        - fusion_rank
      properties:
        memory_id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        content:
          type: string
          example: "User prefers concise, bullet-point responses"
        temporal_level:
          $ref: '#/components/schemas/TemporalLevel'
        salience:
          type: number
          format: float
          description: Effective salience score
          example: 0.92
        outcome_score:
          type: number
          format: float
          description: Contribution from outcome learning
          example: 0.15
        recency_score:
          type: number
          format: float
          description: Temporal recency score
          example: 0.3
        vector_similarity:
          type: number
          format: float
          description: Vector similarity score
          example: 0.87
        fusion_rank:
          type: integer
          description: Final rank after RRF fusion
          example: 1
...
  responses:
    BadRequest:
      description: Bad request - malformed input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "BAD_REQUEST"
            message: "Request body is not valid JSON"
            request_id: "req_abc123"
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "UNAUTHORIZED"
            message: "Invalid or expired authentication token"
            request_id: "req_abc123"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "NOT_FOUND"
            message: "The requested resource was not found"
            request_id: "req_abc123"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            code: "VALIDATION_ERROR"
            message: "Request validation failed"
            request_id: "req_abc123"
            errors:
              - field: "query"
                message: "Query must not be empty"
                code: "required"
    RateLimited:
      description: Too many requests
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per minute
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when limit resets
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "RATE_LIMITED"
            message: "Too many requests. Please retry after 30 seconds."
            request_id: "req_abc123"
            details:
              retry_after: 30
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "INTERNAL_ERROR"
            message: "An unexpected error occurred. Please try again."
            request_id: "req_abc123"